pyproject.toml
[project]
name = "seo-booster"
version = "0.1.0"
description = "SEO content pipeline for AI memory knowledge hub"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "jinja2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
seo-booster = "scripts.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
